Split up components and test popover
