+++ packages/components/src/Calendar/Calendar.test.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import { qantas as theme } from '@roo-ui/themes';
+import { shallowWithTheme } from '@roo-ui/test-utils';
+import { axe } from 'jest-axe';
+
+import Calendar from '.';
+
+describe('<Calendar />', () => {
+  let wrapper;
+
+  beforeEach(() => {
+    wrapper = shallowWithTheme(<Calendar />, theme);
+  });
+
+  it('renders correctly', () => {
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('has no accessibility errors', async () => {
+    expect(await axe(wrapper.html())).toHaveNoViolations();
+  });
+});
