+++ packages/components/src/Calendar/index.js
@@ -1,0 +1,1 @@
+export { default } from './Calendar';
