+++ packages/components/src/index.js
@@ -35,4 +35,5 @@ export { default as Autocomplete } from './Autocomplete';
 export { default as Dropdown } from './Dropdown';
 export { default as Datepicker } from './Datepicker';
 export { default as Popover } from './Popover';
+export { default as Calendar } from './Calendar';
